Flower and Hotspot GameObject Changes
Created data classes for both flowers and hotspots. One version of each object is used to store information about the object. The other is used to create the actual GameObjects within the scene. In the data classes, the method createGameObject(scene) is called on a data object to place it into the scene.